use switch case in InputUI, create atomic Button elements for parsing input values
